ProviderMap: Remove external process reference of ContentProviderRecord
ContentProviderRecords is removed from provider map by activity manager service
when the process hosting the content providers terminates and this causes GC
to free the content provider records but the reference to external process
using the content provider isn't cleared. Fix is for ProviderMap to clear the
reference and unlink for death notification.

Change-Id: I67fd10cee57b02257792d0c9b4569fe11dea7648