Fix the creation of empty file when screenshot is failed.
The doInBackground function does not check the return value from
compress method in Bitmap class. So an empty file is created when
compress fails.

The original code does not clean up file and DB when it fails to
create a screenshot.

This fix adds check the return value and cleans up file and DB
when compress method fails.

Change-Id: I80182a1cba8dde74a11a84e718a8a93770b1597b