Introduce lookup variables for AudioFormat defaults.
Introduce new hidden public variables of what audio format, channel
out config and channel in config to use when *_DEFAULT is selected.
This lookup is later done in AudioRecord and AudioTrack.
Added handling of defaults in AudioRecord and AudioTrack
getMinBufferSize which earlier returned an error instead the buffer
size needed for the specified default(s).

Change-Id: Iaa96375838a4cf67faee61843310995e987f2d3b