Telephony: Introduce config item for Hot Swap support
When a Uicc Card is inserted or removed, Android pops up a prompt
to the user to reboot the device. This prompt is unnecessary if the
modem supports Hot Swapping of cards.

Introduce a new config item to indicate if the modem supports
hotswapping. Frameworks can then check this before propmting the
user.

Bug: 9661862
Change-Id: I0f2c3134525aa2301fd667d0a1819f4419eecae0