Updated Croatian translation of fields in contact entries: phone, address, email, IM types.
The Croatian translation of adjectives used to describe address, phone, email, IM, and other fields in a contact entry was corrected: the translation of "Home" (as in "home number" or "home address") has been corrected from "Početna" ("starting" feminine) to "Privatni", "Privatna", "Privatno"  as is common and found eg. at the Gmail website in Croatian.

Change-Id: I7d49516d7fe5b110ec51c96c22857fadb6f0423f
Signed-off-by: Tomislav Ivek <tomislav.ivek@gmail.com>